+++ src/Export-RecipientPermissions.ps1
@@ -1,3 +1,118 @@
+<#
+.SYNOPSIS
+Export-RecipientPermissions XXXVersionStringXXX
+Document Exchange mailbox access rights, folder permissions, "send as", "send on behalf" and "managed by".
+.DESCRIPTION
+Finds all recipients with a primary SMTP address in an on on-prem or online Exchange environment and documents their
+- mailbox access rights,
+- mailbox folder permissions,
+- "send as" permissions,
+- "send on behalf" permissions,
+- and "managed by" permissions
+
+Easens the move to the cloud, as permission dependencies beyond the supported cross-premises permissions (https://docs.microsoft.com/en-us/Exchange/permissions) can easily be identified and even be represented graphically (sample code included).
+
+Compare exports from different times to detect permission changes (sample code included).
+.LINK
+Github: https://github.com/GruberMarkus/Export-RecipientPermissions
+.PARAMETER ExportFromOnPrem
+Export from On-Prem or from Exchange Online
+$true for export from on-prem, $false for export from Exchange Online
+Default: $false
+.PARAMETER ExchangeConnectionUriList
+Server URIs to connect to
+For on-prem installations, list all Exchange Server Remote PowerShell URIs the script can use
+For Exchange Online use 'https://outlook.office365.com/powershell-liveid/', or the URI specific to your cloud environment
+.PARAMETER ExchangeCredentialUsernameFile, ExchangeCredentialPasswordFile
+Credentials for Exchange connection
+Username and password are stored as encrypted secure strings
+.PARAMETER ParallelJobsExchange, ParallelJobsAD, ParallelJobsLocal
+Maximum Exchange, AD and local sessions/jobs running in parallel
+Watch CPU and RAM usage, and your Exchange throttling policy
+.PARAMETER GrantorFilter
+Grantors to consider
+Only checks recipients that match the filter criteria. Only reduces the number of grantors, not the number of trustees.
+Attributes that can filtered:
+- .DistinguishedName
+- .RecipientType, .RecipientTypeDetails
+- .DisplayName
+- .PrimarySmtpAddress: .Local, .Domain, .Address
+- .EmailAddresses: .PrefixString, .IsPrimaryAddress, .SmtpAddress, .ProxyAddressString
+- On-prem only: .Identity: .tostring() (CN), .DomainId, .Parent (parent CN)
+Set to $null or '' to define all recipients as grantors to consider
+Example: " `$Recipient.primarysmtpaddress.domain -ieq 'example.com'" },
+.PARAMETER ExportMailboxAccessRights
+Rights set on the mailbox itself, such as "FullAccess" and "ReadAccess"
+Default: $true
+.PARAMETER ExportMailboxAccessRightsSelf
+Report mailbox access rights granted to the SID "S-1-5-10" ("NT AUTHORITY\SELF" in English, "NT-AUTORITÄT\SELBST in German, etc.)
+Default: $false
+.PARAMETER ExportMailboxAccessRightsInherited
+Report inherited mailbox access rights (only works on-prem)
+Default: $false
+.PARAMETER ExportMailboxFolderPermissions
+This part of the report can take very long
+Default: $true
+.PARAMETER ExportMailboxFolderPermissionsAnonymous
+Report mailbox folder permissions granted to the special "Anonymous" user ("Anonymous" in English, "Anonym" in German, etc.)
+Default: $false
+.PARAMETER ExportMailboxFolderPermissionsDefault
+Report mailbox folder permissions granted to the special "Default" user ("Default" in English, "Standard" in German, etc.)
+Default: $false
+.PARAMETER ExportMailboxFolderPermissionsOwnerAtLocal
+Exchange Online only. For group mailboxes, export permissions granted to the special "Owner@Local" user.
+Default: $false
+.PARAMETER ExportMailboxFolderPermissionsMemberAtLocal
+Exchange Online only. For group mailboxes, export permissions granted to the special "Member@Local" user.
+Default: $false
+.PARAMETER ExportMailboxFolderPermissionsExcludeFoldertypes
+List of Foldertypes to ignore.
+
+Some known folder types are: Audits, Calendar, CalendarLogging, CommunicatorHistory, Conflicts, Contacts, ConversationActions, DeletedItems, Drafts, ExternalContacts, Files, GalContacts, ImContactList, Inbox, Journal, JunkEmail, LocalFailures, Notes, Outbox, QuickContacts, RecipientCache, RecoverableItemsDeletions, RecoverableItemsPurges, RecoverableItemsRoot, RecoverableItemsVersions, Root, RssSubscription, SentItems, ServerFailures, SyncIssues, Tasks, WorkingSet, YammerFeeds, YammerInbound, YammerOutbound, YammerRoot
+
+Default: 'audits'
+.PARAMETER ExportSendAs
+Export Send As permissions
+Default: $true
+.PARAMETER ExportSendAsSelf
+Export Send As right granted to the SID "S-1-5-10" ("NT AUTHORITY\SELF" in English, "NT-AUTORITÄT\SELBST in German, etc.)
+Default: $false
+.PARAMETER ExportSendOnBehalf
+Export Send On Behalf permissions
+Default: $true
+.PARAMETER ExportManagedBy
+Only for distribution groups, and not to be confused with the "Manager" attribute
+Default: $true
+.PARAMETER ExportFile
+Name (and path) of the permission report file
+Default: '.\export\Export-RecipientPermissions_Result.csv'
+.PARAMETER ErrorFile
+Name (and path) of the error log file
+Set to $null or '' to disable debugging
+Default: '.\export\Export-RecipientPermissions_Error.csv',
+.PARAMETER DebugFile
+Name (and path) of the debug log file
+Set to $null or '' to disable debugging
+Default: ''
+.PARAMETER UpdateInverval
+Interval to update the job progress
+Updates are based von recipients done, not on duration
+Number must be 1 or higher, low numbers mean bigger debug files
+Default: 100
+.INPUTS
+None. You cannot pipe objects to Export-RecipientPermissions.
+.OUTPUTS
+Export-RecipientPermissions writes the current activities, warnings and error messages to the standard output stream.
+.EXAMPLE
+Run Export-RecipientPermissions with default values
+PS> .\Export-RecipientPermissions.ps1
+.NOTES
+Script : Export-RecipientPermissions
+Version: XXXVersionStringXXX
+Web    : https://github.com/GruberMarkus/Export-RecipientPermissions
+License: MIT license (see '.\docs\LICENSE.txt' for details and copyright)
+#>
+
 [CmdletBinding(PositionalBinding = $false)]
 
 
@@ -45,7 +160,7 @@ Param(
     # Mailbox Access Rights
     # Rights set on the mailbox itself, such as "FullAccess" and "ReadAccess"
     [boolean]$ExportMailboxAccessRights = $true,
-    [boolean]$ExportMailboxAccessRightsSelf = $false, # Report mailbox access rights granted to the SID "S-1-5-10" ("NT AUTHORITY\SELF" in English, "NT-AUTORITÄT\SELBST in German, etc.)
+    [boolean]$ExportMailboxAccessRightsSelf = $false, # Report mailbox access rights granted to the SID "S-1-5-10" ("NT AUTHORITY\SELF" in English, "NT-AUTORITÃ„T\SELBST in German, etc.)
     [boolean]$ExportMailboxAccessRightsInherited = $false, # Report inherited mailbox access rights (only works on-prem)
     #
     # Mailbox Folder Permissions
@@ -59,7 +174,7 @@ Param(
     #
     # Send As
     [boolean]$ExportSendAs = $true,
-    [boolean]$ExportSendAsSelf = $false, # Report Send As right granted to the SID "S-1-5-10" ("NT AUTHORITY\SELF" in English, "NT-AUTORITÄT\SELBST in German, etc.)
+    [boolean]$ExportSendAsSelf = $false, # Report Send As right granted to the SID "S-1-5-10" ("NT AUTHORITY\SELF" in English, "NT-AUTORITÃ„T\SELBST in German, etc.)
     #
     # Send On Behalf
     [boolean]$ExportSendOnBehalf = $true,
@@ -101,7 +216,7 @@ $ConnectExchangeOnline = {
         try {
             if (-not (Get-PSSession | Where-Object { ($_.name -like 'ExchangeSession') -and ($_.state -like 'opened') })) {
                 $ExchangeSession = New-PSSession -ConfigurationName Microsoft.Exchange -ConnectionUri $connectionUri -Credential $ExchangeCredential -Authentication Basic -AllowRedirection -Name 'ExchangeSession' -ErrorAction Stop
-                $null = Invoke-Command -Session $ExchangeSession -ScriptBlock { Get-Recipient -ResultSize 1 -wa silentlycontinue -ea stop }
+                $null = Invoke-Command -Session $ExchangeSession -ScriptBlock { Get-Recipient -ResultSize 1 -wa silentlycontinue -ErrorAction Stop } -ErrorAction Stop
             }
             $Stoploop = $true
         } catch {
@@ -121,14 +236,20 @@ $ConnectExchangeOnline = {
 try {
     Set-Location $PSScriptRoot
 
+    if ($PSVersionTable.PSEdition -ieq 'desktop') {
+        $UTF8Encoding = 'UTF8'
+    } else {
+        $UTF8Encoding = 'UTF8BOM'
+    }
+
     if ($ExportFile) {
         $ExportFile = $ExecutionContext.SessionState.Path.GetUnresolvedProviderPathFromPSPath($ExportFile)
     }
-    
+
     if ($ErrorFile) {
         $ErrorFile = $ExecutionContext.SessionState.Path.GetUnresolvedProviderPathFromPSPath($ErrorFile)
     }
-    
+
     if ($DebugFile) {
         $DebugFile = $ExecutionContext.SessionState.Path.GetUnresolvedProviderPathFromPSPath($DebugFile)
     }
@@ -144,8 +265,30 @@ try {
         $null = Start-Transcript -Path $DebugFile -Force
     }
 
+
     Clear-Host
     Write-Host "Start script @$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')@"
+
+
+    Write-Host
+    Write-Host "Script notes @$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')@"
+    Write-Host '  Script : Export-RecipientPermissions'
+    Write-Host '  Version: XXXVersionStringXXX'
+    Write-Host '  Web    : https://github.com/GruberMarkus/Export-RecipientPermissions'
+    Write-Host "  License: MIT license (see '.\docs\LICENSE.txt' for details and copyright)"
+
+
+    Write-Host
+    Write-Host "Script environment and parameters @$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')@"
+    Write-Host "  PowerShell: '$((($($PSVersionTable.PSVersion), $($PSVersionTable.PSEdition), $($PSVersionTable.Platform), $($PSVersionTable.OS)) | Where-Object {$_}) -join "', '")'"
+    Write-Host "  PowerShell bitness: $(if ([Environment]::Is64BitProcess -eq $false) {'Non-'})64-bit process on a $(if ([Environment]::Is64OperatingSystem -eq $false) {'Non-'})64-bit operating system"
+    Write-Host "  Script path: '$PSCommandPath'"
+    Write-Host "  PowerShell invocation: '$(($MyInvocation.Line).trimend([environment]::NewLine))'"
+    Write-Host '  Parameters'
+    foreach ($parameter in (Get-Command -Name $PSCommandPath).Parameters.keys) {
+        Write-Host "    $($parameter): '$((Get-Variable -Name $parameter -EA SilentlyContinue).Value -join ', ')'"
+    }
+
 
     if ($ErrorFile) {
         if (Test-Path $ErrorFile) {
@@ -157,14 +300,13 @@ try {
             }
         } catch {}
         $null = New-Item $ErrorFile -Force
-        '"Timestamp";"Task";"TaskDetail";"Error"' | Out-File $ErrorFile -Encoding utf8bom -Force
+        '"Timestamp";"Task";"TaskDetail";"Error"' | Out-File $ErrorFile -Encoding $UTF8Encoding -Force
     }
 
     if ($DebugFile) {
         foreach ($JobDebugFile in (Get-ChildItem ([io.path]::ChangeExtension(($DebugFile), ('TEMP.*.txt'))))) {
             Remove-Item $JobDebugFile -Force
         }
-        $null = New-Item $DebugFile -Force
     }
 
     if ($ExportFile) {
@@ -177,6 +319,7 @@ try {
             }
         } catch {}
         $null = New-Item $ExportFile -Force
+        '"Grantor Primary SMTP";"Grantor Display Name";"Grantor Recipient Type";"Grantor Environment";"Folder";"Permission";"Allow/Deny";"Inherited";"InheritanceType";"Trustee Original Identity";"Trustee Primary SMTP";"Trustee Display Name";"Trustee Recipient Type";"Trustee Environment"' | Out-File $ExportFile -Encoding $UTF8Encoding -Force
     }
 
     $tempConnectionUriQueue = [System.Collections.Queue]::Synchronized([System.Collections.Queue]::new(10000))
@@ -193,15 +336,15 @@ try {
     if (-not ((Test-Path $ExchangeCredentialUsernameFile) -and (Test-Path $ExchangeCredentialPasswordFile))) {
         Write-Host '  No stored credential found'
         Write-Host '    Username and password are stored as encrypted secure strings'
-        Read-Host -Prompt '    Please enter username for later use (characters are masked)' -AsSecureString | ConvertFrom-SecureString | Out-File $ExchangeCredentialUsernameFile -Force -Encoding utf8bom
-        Read-Host -Prompt '    Please enter password for later use (characters are masked)' -AsSecureString | ConvertFrom-SecureString | Out-File $ExchangeCredentialPasswordFile -Force -Encoding utf8bom
+        Read-Host -Prompt '    Please enter username for later use (characters are masked)' -AsSecureString | ConvertFrom-SecureString | Out-File $ExchangeCredentialUsernameFile -Force -Encoding $UTF8Encoding
+        Read-Host -Prompt '    Please enter password for later use (characters are masked)' -AsSecureString | ConvertFrom-SecureString | Out-File $ExchangeCredentialPasswordFile -Force -Encoding $UTF8Encoding
     }
 
     Write-Host '  Loading credentials encrypted as secure strings'
     Write-Host "    Username file: '$ExchangeCredentialUsernameFile'"
     Write-Host "    Password file: '$ExchangeCredentialPasswordFile'"
     Write-Host '  To change username and/or password, delete one or all of the files mentioned above and run the script again'
-    $ExchangeCredential = New-Object -TypeName System.Management.Automation.PSCredential -ArgumentList ([PSCredential]::new('X', (Get-Content $ExchangeCredentialUsernameFile -Encoding utf8bom | ConvertTo-SecureString)).GetNetworkCredential().Password), (Get-Content $ExchangeCredentialPasswordFile -Encoding utf8bom | ConvertTo-SecureString)
+    $ExchangeCredential = New-Object -TypeName System.Management.Automation.PSCredential -ArgumentList ([PSCredential]::new('X', (Get-Content $ExchangeCredentialUsernameFile -Encoding $UTF8Encoding | ConvertTo-SecureString)).GetNetworkCredential().Password), (Get-Content $ExchangeCredentialPasswordFile -Encoding $UTF8Encoding | ConvertTo-SecureString)
 
 
     # Connect to Exchange
@@ -213,7 +356,7 @@ try {
 
     if ($ExportFromOnPrem -eq $true) {
         $ExchangeSession = New-PSSession -ConfigurationName Microsoft.Exchange -ConnectionUri $connectionUri -Credential $ExchangeCredential -Authentication Kerberos
-        Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { Set-AdServerSettings -ViewEntireForest $True }
+        Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { Set-AdServerSettings -ViewEntireForest $True } -ErrorAction Stop
     } else {
         . ([scriptblock]::Create($ConnectExchangeOnline))
     }
@@ -225,7 +368,7 @@ try {
     Write-Host "  Single-thread operation, use connection '$($connectionUri)'"
 
     $AllRecipients = [system.collections.arraylist]::Synchronized([system.collections.arraylist]::new(1000000))
-    $AllRecipients.AddRange(@((Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { Get-Recipient -resultsize unlimited -WarningAction silentlycontinue | Select-Object -Property identity, distinguishedname, recipienttype, recipienttypedetails, displayname, primarysmtpaddress, emailaddresses, managedby, 'UserFriendlyName' }) | Sort-Object { $_.primarysmtpaddress.address }))
+    $AllRecipients.AddRange(@((Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { Get-Recipient -resultsize unlimited -ErrorAction Stop -WarningAction silentlycontinue | Select-Object -Property identity, distinguishedname, recipienttype, recipienttypedetails, displayname, primarysmtpaddress, emailaddresses, managedby, 'UserFriendlyName' } -ErrorAction Stop) | Sort-Object { $_.primarysmtpaddress.address }))
     $AllRecipients.TrimToSize()
     Write-Host ('  {0:0000000} recipients found' -f $($AllRecipients.count))
 
@@ -236,7 +379,7 @@ try {
     if (($ExportFromOnPrem -eq $false) -and ($ExportSendAs -eq $true)) {
         Write-Host "  Single-thread operation, use connection '$($connectionUri)'"
         $AllRecipientsSendas = [system.collections.arraylist]::Synchronized([system.collections.arraylist]::new($AllRecipients.count * 2))
-        $AllRecipientsSendas.AddRange(@(Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { get-recipientpermission -resultsize unlimited -WarningAction silentlycontinue | Select-Object identity, trustee, trusteesidstring, accessrights, accesscontroltype, isinherited, inheritancetype }))
+        $AllRecipientsSendas.AddRange(@(Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { get-recipientpermission -resultsize unlimited -ErrorAction Stop -WarningAction silentlycontinue | Select-Object identity, trustee, trusteesidstring, accessrights, accesscontroltype, isinherited, inheritancetype } -ErrorAction Stop))
         $AllRecipientsSendas.TrimToSize()
         Write-Host ('  {0:0000000} Send As permissions found' -f $($AllRecipientsSendas.count))
     } else {
@@ -251,13 +394,13 @@ try {
         Write-Host "  Single-thread operation, use connection '$($connectionUri)'"
         $AllRecipientsSendonbehalf = [system.collections.arraylist]::Synchronized([system.collections.arraylist]::new($AllRecipients.count * 2))
         Write-Host "  Mailboxes @$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')@"
-        $AllRecipientsSendonbehalf.AddRange(@(Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { get-mailbox -filter 'GrantSendOnBehalfTo -ne $null' -resultsize unlimited -WarningAction silentlycontinue | Select-Object identity, grantsendonbehalfto }))
+        $AllRecipientsSendonbehalf.AddRange(@(Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { get-mailbox -filter 'GrantSendOnBehalfTo -ne $null' -resultsize unlimited -ErrorAction Stop -WarningAction silentlycontinue | Select-Object identity, grantsendonbehalfto } -ErrorAction Stop))
         Write-Host "  Distribution groups @$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')@"
-        $AllRecipientsSendonbehalf.AddRange(@(Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { get-distributiongroup -filter 'GrantSendOnBehalfTo -ne $null' -resultsize unlimited -WarningAction silentlycontinue | Select-Object identity, grantsendonbehalfto }))
+        $AllRecipientsSendonbehalf.AddRange(@(Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { get-distributiongroup -filter 'GrantSendOnBehalfTo -ne $null' -resultsize unlimited -ErrorAction Stop -WarningAction silentlycontinue | Select-Object identity, grantsendonbehalfto } -ErrorAction Stop))
         Write-Host "  Dynamic Distribution Groups @$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')@"
-        $AllRecipientsSendonbehalf.AddRange(@(Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { get-dynamicdistributiongroup -filter 'GrantSendOnBehalfTo -ne $null' -resultsize unlimited -WarningAction silentlycontinue | Select-Object identity, grantsendonbehalfto }))
+        $AllRecipientsSendonbehalf.AddRange(@(Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { get-dynamicdistributiongroup -filter 'GrantSendOnBehalfTo -ne $null' -resultsize unlimited -ErrorAction Stop -WarningAction silentlycontinue | Select-Object identity, grantsendonbehalfto } -ErrorAction Stop))
         Write-Host "  Unified Groups (Microsoft 365 Groups) @$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')@"
-        $AllRecipientsSendonbehalf.AddRange(@(Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { get-unifiedgroup -filter 'GrantSendOnBehalfTo -ne $null' -resultsize unlimited -WarningAction silentlycontinue | Select-Object identity, grantsendonbehalfto }))
+        $AllRecipientsSendonbehalf.AddRange(@(Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { get-unifiedgroup -filter 'GrantSendOnBehalfTo -ne $null' -resultsize unlimited -ErrorAction Stop -WarningAction silentlycontinue | Select-Object identity, grantsendonbehalfto } -ErrorAction Stop))
         $AllRecipientsSendonbehalf.TrimToSize()
         Write-Host ('  {0:0000000} recipients with Send On Behalf permissions found' -f $($AllRecipientsSendonbehalf.count))
     } else {
@@ -330,6 +473,8 @@ try {
                             $tempQueue,
                             $AllRecipientsGuidToIndex,
                             $DebugFile,
+                            $ErrorFile,
+                            $UTF8Encoding,
                             $ExportFromOnPrem,
                             $ConnectExchangeOnline,
                             $ExchangeCredential,
@@ -348,7 +493,7 @@ try {
                             $connectionUri = $tempConnectionUriQueue.dequeue()
                             if ($ExportFromOnPrem) {
                                 $ExchangeSession = New-PSSession -ConfigurationName Microsoft.Exchange -ConnectionUri $connectionUri -Credential $ExchangeCredential -Authentication Kerberos
-                                Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { Set-AdServerSettings -ViewEntireForest $True }
+                                Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { Set-AdServerSettings -ViewEntireForest $True } -ErrorAction Stop
                             } else {
                                 . ([scriptblock]::Create($ConnectExchangeOnline))
                             }
@@ -369,18 +514,18 @@ try {
                                 Write-Host "  $filterstring"
 
                                 if ($filterstring -ne '') {
-                                    foreach ($securityprincipal in @(Invoke-Command -Session $ExchangeSession -ScriptBlock { get-securityprincipal -filter "$($args[0])" -resultsize unlimited -WarningAction silentlycontinue | Select-Object userfriendlyname, guid } -ArgumentList $filterstring)) {
-                                        try { 
+                                    foreach ($securityprincipal in @(Invoke-Command -Session $ExchangeSession -ScriptBlock { get-securityprincipal -filter "$($args[0])" -resultsize unlimited -ErrorAction Stop -WarningAction silentlycontinue | Select-Object userfriendlyname, guid } -ArgumentList $filterstring -ErrorAction Stop)) {
+                                        try {
                                             Write-Host "  '$($securityprincipal.guid.guid)' = '$($securityprincipal.UserFriendlyName)' @$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')@"
                                             ($AllRecipients[$($AllRecipientsGuidToIndex[$($securityprincipal.guid.guid)])]).UserFriendlyName = $securityprincipal.UserFriendlyName
                                         } catch {
-                                            """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""Find each recipient's UserFriendlyNames"";""$($securityprincipcal.UserFriendlyName)"";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Tee-Object -FilePath $ErrorFile -Append
+                                            """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""Find each recipient's UserFriendlyNames"";""$($securityprincipcal.UserFriendlyName)"";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Add-Content -Path $ErrorFile -PassThru -Encoding $UTF8Encoding
                                         }
                                     }
                                 }
                             }
                         } catch {
-                            """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""Find each recipient's UserFriendlyNames"";"""";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Tee-Object -FilePath $ErrorFile -Append
+                            """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""Find each recipient's UserFriendlyNames"";"""";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Add-Content -Path $ErrorFile -PassThru -Encoding $UTF8Encoding
                         } finally {
                             Remove-PSSession $ExchangeSession
                             if ($DebugFile) {
@@ -396,10 +541,12 @@ try {
                         tempQueue                = $tempQueue
                         AllRecipientsGuidToIndex = $AllRecipientsGuidToIndex
                         DebugFile                = ([io.path]::ChangeExtension(($DebugFile), ('TEMP.{0:0000000}.txt' -f $_)))
+                        ErrorFile                = ([io.path]::ChangeExtension(($ErrorFile), ('TEMP.{0:0000000}.txt' -f $_)))
                         ExportFromOnPrem         = $ExportFromOnPrem
                         ExchangeCredential       = $ExchangeCredential
                         ScriptPath               = $PSScriptRoot
                         ConnectExchangeOnline    = $ConnectExchangeOnline
+                        UTF8Encoding             = $UTF8Encoding
                     }
                 )
 
@@ -444,7 +591,7 @@ try {
                 $null = Stop-Transcript
                 Start-Sleep -Seconds 1
                 foreach ($JobDebugFile in (Get-ChildItem ([io.path]::ChangeExtension(($DebugFile), ('TEMP.*.txt'))))) {
-                    Get-Content $JobDebugFile | Out-File $DebugFile -Append -Encoding utf8bom
+                    Get-Content $JobDebugFile | Out-File $DebugFile -Append
                     Remove-Item $JobDebugFile -Force
                 }
                 $null = Start-Transcript -Path $DebugFile -Append -Force
@@ -452,7 +599,7 @@ try {
 
             if ($ErrorFile) {
                 foreach ($JobErrorFile in (Get-ChildItem ([io.path]::ChangeExtension(($ErrorFile), ('TEMP.*.txt'))))) {
-                    Get-Content $JobErrorFile | Out-File $ErrorFile -Append -Encoding utf8bom
+                    Get-Content $JobErrorFile | Out-File $ErrorFile -Append -Encoding $UTF8Encoding
                     Remove-Item $JobErrorFile -Force
                 }
             }
@@ -537,6 +684,7 @@ try {
                             $ExportMailboxAccessRightsInherited,
                             $ExportFile,
                             $ErrorFile,
+                            $UTF8Encoding,
                             $AllRecipientsUfnToIndex,
                             $DebugFile,
                             $ExportFromOnPrem,
@@ -557,7 +705,7 @@ try {
                             $connectionUri = $tempConnectionUriQueue.dequeue()
                             if ($ExportFromOnPrem) {
                                 $ExchangeSession = New-PSSession -ConfigurationName Microsoft.Exchange -ConnectionUri $connectionUri -Credential $ExchangeCredential -Authentication Kerberos
-                                Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { Set-AdServerSettings -ViewEntireForest $True }
+                                Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { Set-AdServerSettings -ViewEntireForest $True } -ErrorAction Stop
                             } else {
                                 . ([scriptblock]::Create($ConnectExchangeOnline))
                             }
@@ -591,10 +739,10 @@ try {
                                     foreach ($MailboxPermission in
                                         @($(
                                                 if ($ExportFromOnPrem) {
-                                                    Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { get-mailboxpermission -identity $args[0] -resultsize unlimited -WarningAction silentlycontinue | Select-Object -Property identity, user, accessrights, deny, isinherited, inheritanceType } -ArgumentList $GrantorPrimarySMTP | Select-Object identity, user, accessrights, deny, isinherited, inheritanceType
+                                                    Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { get-mailboxpermission -identity $args[0] -resultsize unlimited -ErrorAction Stop -WarningAction silentlycontinue | Select-Object -Property identity, user, accessrights, deny, isinherited, inheritanceType } -ArgumentList $GrantorPrimarySMTP  -ErrorAction Stop | Select-Object identity, user, accessrights, deny, isinherited, inheritanceType
                                                 } else {
                                                     if ($GrantorRecipientTypeDetails -ine 'groupmailbox') {
-                                                        Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { get-mailboxpermission -identity $args[0] -resultsize unlimited -WarningAction silentlycontinue | Select-Object -Property identity, user, accessrights, deny, isinherited, inheritanceType } -ArgumentList $GrantorPrimarySMTP | Select-Object identity, user, accessrights, deny, isinherited, inheritanceType
+                                                        Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { get-mailboxpermission -identity $args[0] -resultsize unlimited -ErrorAction Stop -WarningAction silentlycontinue | Select-Object -Property identity, user, accessrights, deny, isinherited, inheritanceType } -ArgumentList $GrantorPrimarySMTP  -ErrorAction Stop | Select-Object identity, user, accessrights, deny, isinherited, inheritanceType
                                                     }
                                                 }
                                             ))
@@ -637,12 +785,12 @@ try {
                                         }
                                     }
                                 } catch {
-                                    """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""Find each recipient's UserFriendlyNames"";""$($GrantorPrimarySMTP)"";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Tee-Object -FilePath $ErrorFile -Append
+                                    """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""Find each recipient's UserFriendlyNames"";""$($GrantorPrimarySMTP)"";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Add-Content -Path $ErrorFile -PassThru -Encoding $UTF8Encoding
                                 }
-                                $ExportFileResult | Sort-Object -Unique | Out-File ([io.path]::ChangeExtension(($ExportFile), ('TEMP.{0:0000000}.txt' -f $RecipientID))) -Append -Encoding utf8bom -Force
+                                $ExportFileResult | Sort-Object -Unique | Out-File ([io.path]::ChangeExtension(($ExportFile), ('TEMP.{0:0000000}.txt' -f $RecipientID))) -Append -Force -Encoding $UTF8Encoding
                             }
                         } catch {
-                            """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""Mailbox Access Rights"";"""";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Tee-Object -FilePath $ErrorFile -Append
+                            """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""Mailbox Access Rights"";"""";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Add-Content -Path $ErrorFile -PassThru -Encoding $UTF8Encoding
                         } finally {
                             Remove-PSSession $ExchangeSession
 
@@ -667,6 +815,7 @@ try {
                         ExchangeCredential                 = $ExchangeCredential
                         ScriptPath                         = $PSScriptRoot
                         ConnectExchangeOnline              = $ConnectExchangeOnline
+                        UTF8Encoding                       = $UTF8Encoding
                     }
                 )
 
@@ -711,7 +860,7 @@ try {
                 $null = Stop-Transcript
                 Start-Sleep -Seconds 1
                 foreach ($JobDebugFile in (Get-ChildItem ([io.path]::ChangeExtension(($DebugFile), ('TEMP.*.txt'))))) {
-                    Get-Content $JobDebugFile | Out-File $DebugFile -Append -Encoding utf8bom
+                    Get-Content $JobDebugFile | Out-File $DebugFile -Append
                     Remove-Item $JobDebugFile -Force
                 }
                 $null = Start-Transcript -Path $DebugFile -Append -Force
@@ -719,7 +868,7 @@ try {
 
             if ($ErrorFile) {
                 foreach ($JobErrorFile in (Get-ChildItem ([io.path]::ChangeExtension(($ErrorFile), ('TEMP.*.txt'))))) {
-                    Get-Content $JobErrorFile | Out-File $ErrorFile -Append -Encoding utf8bom
+                    Get-Content $JobErrorFile | Out-File $ErrorFile -Append -Encoding $UTF8Encoding
                     Remove-Item $JobErrorFile -Force
                 }
             }
@@ -770,8 +919,10 @@ try {
                             $ExportMailboxFolderPermissionsDefault,
                             $ExportMailboxFolderPermissionsOwnerAtLocal,
                             $ExportMailboxFolderPermissionsMemberAtLocal,
+                            $ExportMailboxFolderPermissionsExcludeFoldertypes,
                             $ExportFile,
                             $ErrorFile,
+                            $UTF8Encoding,
                             $AllRecipientsGuidToIndex,
                             $DebugFile,
                             $ExportFromOnPrem,
@@ -791,7 +942,7 @@ try {
                             $connectionUri = $tempConnectionUriQueue.dequeue()
                             if ($ExportFromOnPrem) {
                                 $ExchangeSession = New-PSSession -ConfigurationName Microsoft.Exchange -ConnectionUri $connectionUri -Credential $ExchangeCredential -Authentication Kerberos
-                                Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { Set-AdServerSettings -ViewEntireForest $True }
+                                Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { Set-AdServerSettings -ViewEntireForest $True } -ErrorAction Stop
                             } else {
                                 . ([scriptblock]::Create($ConnectExchangeOnline))
                             }
@@ -816,10 +967,9 @@ try {
 
                                 Write-Host "$($GrantorPrimarySMTP), $($GrantorRecipientType)/$($GrantorRecipientTypeDetails) @$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')@"
 
-                                $Folders = Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { get-mailboxfolderstatistics -identity $args[0] | Select-Object folderid, folderpath, foldertype } -ArgumentList $GrantorPrimarySMTP
-
-                                try {
-                                    foreach ($Folder in $Folders) {
+                                $Folders = Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { get-mailboxfolderstatistics -identity $args[0] -ErrorAction Stop -WarningAction silentlycontinue | Select-Object folderid, folderpath, foldertype } -ArgumentList $GrantorPrimarySMTP -ErrorAction Stop
+                                foreach ($Folder in $Folders) {
+                                    try {
                                         if ($folder.foldertype -iin $ExportMailboxFolderPermissionsExcludeFoldertypes) { continue }
 
                                         if ($Folder.foldertype -ieq 'root') { $Folder.folderpath = '/' }
@@ -828,12 +978,12 @@ try {
                                         foreach ($FolderPermissions in
                                             @($(
                                                     if ($ExportFromOnPrem) {
-                                                        (Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { get-mailboxfolderpermission -identity $args[0] -WarningAction silentlycontinue | Select-Object identity, user, accessrights } -ArgumentList "$($GrantorPrimarySMTP):$($Folder.folderid)")
+                                                        (Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { get-mailboxfolderpermission -identity $args[0] -ErrorAction stop -WarningAction silentlycontinue | Select-Object identity, user, accessrights } -ArgumentList "$($GrantorPrimarySMTP):$($Folder.folderid)" -ErrorAction Stop)
                                                     } else {
                                                         if ($GrantorRecipientTypeDetails -ieq 'groupmailbox') {
-                                                            (Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { get-mailboxfolderpermission -identity $args[0] -groupmailbox -WarningAction silentlycontinue | Select-Object identity, user, accessrights } -ArgumentList "$($GrantorPrimarySMTP):$($Folder.folderid)")
+                                                            (Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { get-mailboxfolderpermission -identity $args[0] -groupmailbox -ErrorAction stop -WarningAction silentlycontinue | Select-Object identity, user, accessrights } -ArgumentList "$($GrantorPrimarySMTP):$($Folder.folderid)" -ErrorAction Stop)
                                                         } else {
-                                                            (Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { get-mailboxfolderpermission -identity $args[0] -WarningAction silentlycontinue | Select-Object identity, user, accessrights } -ArgumentList "$($GrantorPrimarySMTP):$($Folder.folderid)")
+                                                            (Invoke-Command -Session $ExchangeSession -HideComputerName -ScriptBlock { get-mailboxfolderpermission -identity $args[0] -ErrorAction stop -WarningAction silentlycontinue | Select-Object identity, user, accessrights } -ArgumentList "$($GrantorPrimarySMTP):$($Folder.folderid)" -ErrorAction Stop)
                                                         }
                                                     }
                                                 ))
@@ -893,14 +1043,14 @@ try {
                                                 }
                                             }
                                         }
+                                    } catch {
+                                        """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""Mailbox Folder permissions"";""$($GrantorPrimarySMTP):$($Folder.folderpath)"";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Add-Content -Path $ErrorFile -PassThru -Encoding $UTF8Encoding
                                     }
-                                } catch {
-                                    """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""Mailbox Folder permissions"";""$($GrantorPrimarySMTP):\$($Folder.folderpath)"";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Tee-Object -FilePath $ErrorFile -Append
                                 }
-                                $ExportFileResult | Sort-Object -Unique | Out-File ([io.path]::ChangeExtension(($ExportFile), ('TEMP.{0:0000000}.txt' -f $RecipientID))) -Append -Encoding utf8bom -Force
+                                $ExportFileResult | Sort-Object -Unique | Out-File ([io.path]::ChangeExtension(($ExportFile), ('TEMP.{0:0000000}.txt' -f $RecipientID))) -Append -Force -Encoding $UTF8Encoding
                             }
                         } catch {
-                            """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""Mailbox Folder permissions"";"""";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Tee-Object -FilePath $ErrorFile -Append
+                            """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""Mailbox Folder permissions"";"""";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Add-Content -Path $ErrorFile -PassThru -Encoding $UTF8Encoding
 
                         } finally {
                             Remove-PSSession $ExchangeSession
@@ -912,22 +1062,24 @@ try {
                     }
                 ).AddParameters(
                     @{
-                        AllRecipients                               = $AllRecipients
-                        tempConnectionUriQueue                      = $tempConnectionUriQueue
-                        tempQueue                                   = $tempQueue
-                        ExportMailboxFolderPermissions              = $ExportMailboxFolderPermissions
-                        ExportMailboxFolderPermissionsAnonymous     = $ExportMailboxFolderPermissionsAnonymous
-                        ExportMailboxFolderPermissionsDefault       = $ExportMailboxFolderPermissionsDefault
-                        ExportMailboxFolderPermissionsOwnerAtLocal  = $ExportMailboxFolderPermissionsOwnerAtLocal
-                        ExportMailboxFolderPermissionsMemberAtLocal = $ExportMailboxFolderPermissionsMemberAtLocal
-                        ExportFile                                  = $ExportFile
-                        AllRecipientsGuidToIndex                    = $AllRecipientsGuidToIndex
-                        ErrorFile                                   = ([io.path]::ChangeExtension(($ErrorFile), ('TEMP.{0:0000000}.txt' -f $_)))
-                        DebugFile                                   = ([io.path]::ChangeExtension(($DebugFile), ('TEMP.{0:0000000}.txt' -f $_)))
-                        ExportFromOnPrem                            = $ExportFromOnPrem
-                        ExchangeCredential                          = $ExchangeCredential
-                        ScriptPath                                  = $PSScriptRoot
-                        ConnectExchangeOnline                       = $ConnectExchangeOnline
+                        AllRecipients                                    = $AllRecipients
+                        tempConnectionUriQueue                           = $tempConnectionUriQueue
+                        tempQueue                                        = $tempQueue
+                        ExportMailboxFolderPermissions                   = $ExportMailboxFolderPermissions
+                        ExportMailboxFolderPermissionsAnonymous          = $ExportMailboxFolderPermissionsAnonymous
+                        ExportMailboxFolderPermissionsDefault            = $ExportMailboxFolderPermissionsDefault
+                        ExportMailboxFolderPermissionsOwnerAtLocal       = $ExportMailboxFolderPermissionsOwnerAtLocal
+                        ExportMailboxFolderPermissionsMemberAtLocal      = $ExportMailboxFolderPermissionsMemberAtLocal
+                        ExportMailboxFolderPermissionsExcludeFoldertypes = $ExportMailboxFolderPermissionsExcludeFoldertypes
+                        ExportFile                                       = $ExportFile
+                        AllRecipientsGuidToIndex                         = $AllRecipientsGuidToIndex
+                        ErrorFile                                        = ([io.path]::ChangeExtension(($ErrorFile), ('TEMP.{0:0000000}.txt' -f $_)))
+                        DebugFile                                        = ([io.path]::ChangeExtension(($DebugFile), ('TEMP.{0:0000000}.txt' -f $_)))
+                        ExportFromOnPrem                                 = $ExportFromOnPrem
+                        ExchangeCredential                               = $ExchangeCredential
+                        ScriptPath                                       = $PSScriptRoot
+                        ConnectExchangeOnline                            = $ConnectExchangeOnline
+                        UTF8Encoding                                     = $UTF8Encoding
                     }
                 )
 
@@ -972,7 +1124,7 @@ try {
                 $null = Stop-Transcript
                 Start-Sleep -Seconds 1
                 foreach ($JobDebugFile in (Get-ChildItem ([io.path]::ChangeExtension(($DebugFile), ('TEMP.*.txt'))))) {
-                    Get-Content $JobDebugFile | Out-File $DebugFile -Append -Encoding utf8bom
+                    Get-Content $JobDebugFile | Out-File $DebugFile -Append
                     Remove-Item $JobDebugFile -Force
                 }
                 $null = Start-Transcript -Path $DebugFile -Append -Force
@@ -980,7 +1132,7 @@ try {
 
             if ($ErrorFile) {
                 foreach ($JobErrorFile in (Get-ChildItem ([io.path]::ChangeExtension(($ErrorFile), ('TEMP.*.txt'))))) {
-                    Get-Content $JobErrorFile | Out-File $ErrorFile -Append -Encoding utf8bom
+                    Get-Content $JobErrorFile | Out-File $ErrorFile -Append -Encoding $UTF8Encoding
                     Remove-Item $JobErrorFile -Force
                 }
             }
@@ -1029,6 +1181,7 @@ try {
                             $tempQueue,
                             $ExportFile,
                             $ErrorFile,
+                            $UTF8Encoding,
                             $AllRecipientsUfnToIndex,
                             $AllRecipientsSmtpToIndex,
                             $ExportSendAsSelf,
@@ -1160,12 +1313,12 @@ try {
                                         }
                                     }
                                 } catch {
-                                    """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""SendAs"";""$($GrantorPrimarySMTP)"";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Tee-Object -FilePath $ErrorFile -Append
+                                    """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""SendAs"";""$($GrantorPrimarySMTP)"";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Add-Content -Path $ErrorFile -PassThru -Encoding $UTF8Encoding
                                 }
                             }
-                            $ExportFileResult | Sort-Object -Unique | Out-File ([io.path]::ChangeExtension(($ExportFile), ('TEMP.{0:0000000}.txt' -f $RecipientID))) -Append -Encoding utf8bom -Force
+                            $ExportFileResult | Sort-Object -Unique | Out-File ([io.path]::ChangeExtension(($ExportFile), ('TEMP.{0:0000000}.txt' -f $RecipientID))) -Append -Force -Encoding $UTF8Encoding
                         } catch {
-                            """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""Send As"";"""";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Tee-Object -FilePath $ErrorFile -Append
+                            """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""Send As"";"""";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Add-Content -Path $ErrorFile -PassThru -Encoding $UTF8Encoding
                         } finally {
                             if ($DebugFile) {
                                 $null = Stop-Transcript
@@ -1186,6 +1339,7 @@ try {
                         ExportFromOnPrem         = $ExportFromOnPrem
                         ScriptPath               = $PSScriptRoot
                         AllRecipientsSendas      = $AllRecipientsSendas
+                        UTF8Encoding             = $UTF8Encoding
                     }
                 )
 
@@ -1230,7 +1384,7 @@ try {
                 $null = Stop-Transcript
                 Start-Sleep -Seconds 1
                 foreach ($JobDebugFile in (Get-ChildItem ([io.path]::ChangeExtension(($DebugFile), ('TEMP.*.txt'))))) {
-                    Get-Content $JobDebugFile | Out-File $DebugFile -Append -Encoding utf8bom
+                    Get-Content $JobDebugFile | Out-File $DebugFile -Append
                     Remove-Item $JobDebugFile -Force
                 }
                 $null = Start-Transcript -Path $DebugFile -Append -Force
@@ -1238,7 +1392,7 @@ try {
 
             if ($ErrorFile) {
                 foreach ($JobErrorFile in (Get-ChildItem ([io.path]::ChangeExtension(($ErrorFile), ('TEMP.*.txt'))))) {
-                    Get-Content $JobErrorFile | Out-File $ErrorFile -Append -Encoding utf8bom
+                    Get-Content $JobErrorFile | Out-File $ErrorFile -Append -Encoding $UTF8Encoding
                     Remove-Item $JobErrorFile -Force
                 }
             }
@@ -1289,6 +1443,7 @@ try {
                             $tempQueue,
                             $ExportFile,
                             $ErrorFile,
+                            $UTF8Encoding,
                             $AllRecipientsDnToIndex,
                             $AllRecipientsGuidToIndex,
                             $DebugFile,
@@ -1408,12 +1563,12 @@ try {
                                         }
                                     }
                                 } catch {
-                                    """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""Send On Behalf"";""$($GrantorPrimarySMTP)"";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Tee-Object -FilePath $ErrorFile -Append
+                                    """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""Send On Behalf"";""$($GrantorPrimarySMTP)"";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Add-Content -Path $ErrorFile -PassThru -Encoding $UTF8Encoding
                                 }
-                                $ExportFileResult | Sort-Object -Unique | Out-File ([io.path]::ChangeExtension(($ExportFile), ('TEMP.{0:0000000}.txt' -f $RecipientID))) -Append -Encoding utf8bom -Force
+                                $ExportFileResult | Sort-Object -Unique | Out-File ([io.path]::ChangeExtension(($ExportFile), ('TEMP.{0:0000000}.txt' -f $RecipientID))) -Append -Force -Encoding $UTF8Encoding
                             }
                         } catch {
-                            """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""Send On Behalf"";"""";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Tee-Object -FilePath $ErrorFile -Append
+                            """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""Send On Behalf"";"""";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Add-Content -Path $ErrorFile -PassThru -Encoding $UTF8Encoding
                         } finally {
                             if ($DebugFile) {
                                 $null = Stop-Transcript
@@ -1433,6 +1588,7 @@ try {
                         ExportFromOnPrem          = $ExportFromOnPrem
                         ScriptPath                = $PSScriptRoot
                         AllRecipientsSendonbehalf = $AllRecipientsSendonbehalf
+                        UTF8Encoding              = $UTF8Encoding
                     }
                 )
 
@@ -1477,7 +1633,7 @@ try {
                 $null = Stop-Transcript
                 Start-Sleep -Seconds 1
                 foreach ($JobDebugFile in (Get-ChildItem ([io.path]::ChangeExtension(($DebugFile), ('TEMP.*.txt'))))) {
-                    Get-Content $JobDebugFile | Out-File $DebugFile -Append -Encoding utf8bom
+                    Get-Content $JobDebugFile | Out-File $DebugFile -Append
                     Remove-Item $JobDebugFile -Force
                 }
                 $null = Start-Transcript -Path $DebugFile -Append -Force
@@ -1485,7 +1641,7 @@ try {
 
             if ($ErrorFile) {
                 foreach ($JobErrorFile in (Get-ChildItem ([io.path]::ChangeExtension(($ErrorFile), ('TEMP.*.txt'))))) {
-                    Get-Content $JobErrorFile | Out-File $ErrorFile -Append -Encoding utf8bom
+                    Get-Content $JobErrorFile | Out-File $ErrorFile -Append -Encoding $UTF8Encoding
                     Remove-Item $JobErrorFile -Force
                 }
             }
@@ -1533,6 +1689,7 @@ try {
                             $tempQueue,
                             $ExportFile,
                             $ErrorFile,
+                            $UTF8Encoding,
                             $AllRecipientsGuidToIndex,
                             $DebugFile,
                             $ScriptPath
@@ -1543,6 +1700,7 @@ try {
                             if ($DebugFile) {
                                 $null = Start-Transcript -Path $DebugFile -Force
                             }
+
                             Write-Host "Managed By @$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')@"
 
                             $ExportFileResult = [system.collections.arraylist]::new(1000)
@@ -1602,12 +1760,12 @@ try {
                                         }
                                     }
                                 } catch {
-                                    """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""Managed By"";""$($GrantorPrimarySMTP)"";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Tee-Object -FilePath $ErrorFile -Append
+                                    """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""Managed By"";""$($GrantorPrimarySMTP)"";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Add-Content -Path $ErrorFile -PassThru -Encoding $UTF8Encoding
                                 }
-                                $ExportFileResult | Sort-Object -Unique | Out-File ([io.path]::ChangeExtension(($ExportFile), ('TEMP.{0:0000000}.txt' -f $RecipientID))) -Append -Encoding utf8bom -Force
+                                $ExportFileResult | Sort-Object -Unique | Out-File ([io.path]::ChangeExtension(($ExportFile), ('TEMP.{0:0000000}.txt' -f $RecipientID))) -Append -Force -Encoding $UTF8Encoding
                             }
                         } catch {
-                            """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""Managed By"";""$($GrantorPrimarySMTP)"";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Tee-Object -FilePath $ErrorFile -Append
+                            """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";""Managed By"";""$($GrantorPrimarySMTP)"";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Add-Content -Path $ErrorFile -PassThru -Encoding $UTF8Encoding
                         } finally {
                             if ($DebugFile) {
                                 $null = Stop-Transcript
@@ -1624,6 +1782,7 @@ try {
                         ErrorFile                = ([io.path]::ChangeExtension(($ErrorFile), ('TEMP.{0:0000000}.txt' -f $_)))
                         DebugFile                = ([io.path]::ChangeExtension(($DebugFile), ('TEMP.{0:0000000}.txt' -f $_)))
                         ScriptPath               = $PSScriptRoot
+                        UTF8Encoding             = $UTF8Encoding
                     }
                 )
 
@@ -1668,7 +1827,7 @@ try {
                 $null = Stop-Transcript
                 Start-Sleep -Seconds 1
                 foreach ($JobDebugFile in (Get-ChildItem ([io.path]::ChangeExtension(($DebugFile), ('TEMP.*.txt'))))) {
-                    Get-Content $JobDebugFile | Out-File $DebugFile -Append -Encoding utf8bom
+                    Get-Content $JobDebugFile | Out-File $DebugFile -Append
                     Remove-Item $JobDebugFile -Force
                 }
                 $null = Start-Transcript -Path $DebugFile -Append -Force
@@ -1676,7 +1835,7 @@ try {
 
             if ($ErrorFile) {
                 foreach ($JobErrorFile in (Get-ChildItem ([io.path]::ChangeExtension(($ErrorFile), ('TEMP.*.txt'))))) {
-                    Get-Content $JobErrorFile | Out-File $ErrorFile -Append -Encoding utf8bom
+                    Get-Content $JobErrorFile | Out-File $ErrorFile -Append -Encoding $UTF8Encoding
                     Remove-Item $JobErrorFile -Force
                 }
             }
@@ -1694,22 +1853,23 @@ try {
     Write-Host "Create sorted export file @$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')@"
     Write-Host "  '$ExportFile'"
 
-    '"Grantor Primary SMTP";"Grantor Display Name";"Grantor Recipient Type";"Grantor Environment";"Folder";"Permission";"Allow/Deny";"Inherited";"InheritanceType";"Trustee Original Identity";"Trustee Primary SMTP";"Trustee Display Name";"Trustee Recipient Type";"Trustee Environment"' | Out-File $ExportFile -Encoding utf8bom -Force
-
     foreach ($RecipientFile in (Get-ChildItem ([io.path]::ChangeExtension(($ExportFile), ('TEMP.*.txt'))))) {
-        Get-Content $RecipientFile | Sort-Object -Unique | Out-File $ExportFile -Encoding utf8bom -Append -Force
+        Get-Content $RecipientFile | Sort-Object -Unique | Out-File $ExportFile -Append -Force -Encoding $UTF8Encoding
         Remove-Item $Recipientfile -Force
     }
 
 } catch {
     Write-Host 'Unexpected error. Exiting.'
-    """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";"""";"""";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Tee-Object -FilePath $ErrorFile -Append
+    """$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')"";"""";"""";""$($_ | Out-String)""" -replace '(?<!;|^)"(?!;|$)', '""' | Add-Content -Path $ErrorFile -PassThru -Encoding $UTF8Encoding
 } finally {
+    Write-Host
     Write-Host
 
     if ($ExchangeSession) { Remove-PSSession $ExchangeSession }
 
-    Write-Host 'Closing runspaces and runspacepool, please wait'
+    Write-Host "Clean-up @$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')@"
+    Write-Host "  Runspaces and RunspacePool @$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')@"
+
     if ($runspaces) {
         foreach ($runspace in $runspaces) {
             $runspace.PowerShell.Dispose()
@@ -1719,9 +1879,18 @@ try {
         $RunspacePool.dispose()
     }
 
+    Write-Host "  Temporary result files @$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')@"
     foreach ($RecipientFile in (Get-ChildItem ([io.path]::ChangeExtension(($ExportFile), ('TEMP.*.txt'))))) {
-        Get-Content $RecipientFile | Sort-Object -Unique | Out-File $ExportFile -Encoding utf8bom -Append -Force
+        Get-Content $RecipientFile | Sort-Object -Unique | Out-File $ExportFile -Append -Force -Encoding $UTF8Encoding
         Remove-Item $Recipientfile -Force
+    }
+
+    if ($ErrorFile) {
+        Write-Host "  Temporary error files @$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')@"
+        foreach ($JobErrorFile in (Get-ChildItem ([io.path]::ChangeExtension(($ErrorFile), ('TEMP.*.txt'))))) {
+            Get-Content $JobErrorFile | Out-File $ErrorFile -Append -Encoding $UTF8Encoding
+            Remove-Item $JobErrorFile -Force
+        }
     }
 
     Write-Host
@@ -1729,18 +1898,12 @@ try {
     Write-Host
 
     if ($DebugFile) {
+        Write-Host "  Temporary debug files @$(Get-Date -Format 'yyyy-MM-ddTHH:mm:sszzz')@"
         $null = Stop-Transcript
         Start-Sleep -Seconds 1
         foreach ($JobDebugFile in (Get-ChildItem ([io.path]::ChangeExtension(($DebugFile), ('TEMP.*.txt'))))) {
-            Get-Content $JobDebugFile | Out-File $DebugFile -Append -Encoding utf8bom
+            Get-Content $JobDebugFile | Out-File $DebugFile -Append
             Remove-Item $JobDebugFile -Force
-        }
-    }
-
-    if ($ErrorFile) {
-        foreach ($JobErrorFile in (Get-ChildItem ([io.path]::ChangeExtension(($ErrorFile), ('TEMP.*.txt'))))) {
-            Get-Content $JobErrorFile | Out-File $ErrorFile -Append -Encoding utf8bom
-            Remove-Item $JobErrorFile -Force
         }
     }